tts-service/requirements.txt
fastapi
coqui-tts
uvicorn 
soundfile 
num2words